fix hypot bug, added output and transform for amcl

+++ src/landmark_localization/ll_amcl2d.py
@@ -99,12 +99,13 @@ class AMCL2D(llc.LandmarkLocalization):
             if 'a' in landmark_param:
                 al.append([landmark_param['x'], landmark_param['y'], landmark_param['a'], landmark_param['sa']])
         if len(rl) != 0:
-            rl = np.array(rl)
+            rl = np.array(rl, dtype = np.float64)
             mP = np.tile(self.P, [rl.shape[0], 1])
             mrl = np.repeat(rl, self.P.shape[0], axis = 0)
             dx = mrl[:,0] - mP[:,0]
             dy = mrl[:,1] - mP[:,1]     
             #print(dx.shape, dy.shape)
+            #print(np.__version__)
             dr = np.hypot(dx, dy) - mrl[:,2]
             w_all = norm.pdf(dr, scale = mrl[:,3])            
             w = w_all.reshape((rl.shape[0], self.P.shape[0]))            
@@ -117,7 +118,7 @@ class AMCL2D(llc.LandmarkLocalization):
                 self.W *= w
             
         if len(al) != 0:
-            al = np.array(al)        
+            al = np.array(al, dtype = np.float64)        
             mP = np.tile(self.P, [rl.shape[0], 1])
             mal = np.repeat(al, self.P.shape[0], axis = 0)
             dx = mrl[:,0] - mP[:,0]
